Fixed floating point arithmetic.

+++ src/lib/access/RadixJoin.h
@@ -24,8 +24,8 @@ protected:
   // for determineDynamicCount
   // overridden from PlanOperation
   virtual size_t getTotalTableSize();
-  virtual double calcMinMts(double totalTblSizeIn100k);
-  virtual double calcA(double totalTblSizeIn100k);
+  virtual size_t calcMinMts(size_t totalTblSizeIn100k);
+  virtual size_t calcA(size_t totalTblSizeIn100k);
   virtual double min_mts_a() { return -32.7333223568781 ; }
   virtual double min_mts_b() { return 20.6071622571548 ; }
   virtual double a_a() { return 0.0499042793549051 ; }
